requirements.txt
Flask
Flask-SQLAlchemy
Faker
bcrypt